requirements.txt
neo4j==5.28.1
neo4j-graphrag[openai]==1.7.0
pandas==2.2.3
numpy==1.26.4
python-dotenv==1.1.0